+++ src/segments/fetchAndParse.ts
@@ -5,7 +5,7 @@ import type { Context } from "@utils/context";
 import type { CheerioAPI } from "cheerio";
 import type { ResultAsync } from "neverthrow";
 
-import { fetchAndParse } from "@utils/smarterFetch";
+import { fetchAndParse } from "@utils/smartFetch";
 
 export const fetchAndParseHTML = (
 	ctx: Context,
