+++ src/utils/smartFetch.ts
